DSS-846 Add temp hash unit test files
