+++ packages/coordinator/tests/resources/signed-prescriptions/GenerateHashSignature.xml
@@ -3,7 +3,7 @@
   <code codeSystem="2.16.840.1.113883.2.1.3.2.4.15" code="163501000000109">
     <originalText>Prescription - FocusActOrEvent </originalText>
   </code>
-  <effectiveTime xmlns:hl7="urn:hl7-org:v3" value="20250806125914"/>
+  <effectiveTime xmlns:hl7="urn:hl7-org:v3" value="20250812120729"/>
   <typeId root="2.16.840.1.113883.2.1.3.2.4.18.7" extension="PORX_MT132004UK31"/>
   <recordTarget typeCode="RCT">
     <Patient classCode="PAT">
@@ -45,7 +45,7 @@
     <templateId root="2.16.840.1.113883.2.1.3.2.4.18.2" extension="CSAB_RM-NPfITUK10.pertinentInformation"/>
     <pertinentPrescription classCode="SBADM" moodCode="RQO">
       <id root="E462B682-FF5F-471B-BCEB-DA60D648B7EE"/>
-      <id xmlns:hl7="urn:hl7-org:v3" extension="7E1AE7-9439DE-425B9+" root="2.16.840.1.113883.2.1.3.2.4.18.8"/>
+      <id xmlns:hl7="urn:hl7-org:v3" extension="FC327D-5AA02D-466E8S" root="2.16.840.1.113883.2.1.3.2.4.18.8"/>
       <code codeSystem="2.16.840.1.113883.2.1.3.2.4.15" code="225426007"/>
       <effectiveTime nullFlavor="NA"/>
       <!-- Prescription Nomination Detail -->
@@ -57,14 +57,14 @@
         </AgentOrgSDS>
       </performer>
       <author typeCode="AUT" contextControlCode="OP">
-        <time xmlns:hl7="urn:hl7-org:v3" value="20250806125914"/>
+        <time xmlns:hl7="urn:hl7-org:v3" value="20250812120729"/>
         <signatureText>
           <Signature xmlns="http://www.w3.org/2000/09/xmldsig#">
-            <SignedInfo xmlns="http://www.w3.org/2000/09/xmldsig#"><CanonicalizationMethod Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"></CanonicalizationMethod><SignatureMethod Algorithm="http://www.w3.org/2000/09/xmldsig#rsa-sha1"></SignatureMethod><Reference><Transforms><Transform Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"></Transform></Transforms><DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha1"></DigestMethod><DigestValue>C0ljh8shA++0KSPIVPQfHfK8NBY=</DigestValue></Reference></SignedInfo>
-            <SignatureValue>HImpN+2s7vwOnFHcicAnx5kRsKOYTtQeK0V1bnOk9UIJ7MkzqsVPSrA2TN3SUSIh+aHpCmZbRRgF5qj/gI+4GhK3OaxBxA3V/23BD99cjyYtUK3qrnRM2cr03zDg9I5ZuQL/MdBFMZdSTvUDykJ5Nd6pB0a6bdY2uRXKgPMXh9M=</SignatureValue>
+            <SignedInfo xmlns="http://www.w3.org/2000/09/xmldsig#"><CanonicalizationMethod Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"></CanonicalizationMethod><SignatureMethod Algorithm="http://www.w3.org/2000/09/xmldsig#rsa-sha1"></SignatureMethod><Reference><Transforms><Transform Algorithm="http://www.w3.org/2001/10/xml-exc-c14n#"></Transform></Transforms><DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha1"></DigestMethod><DigestValue>Ht3Z8p1HSTqndqRtmIpd++9PZ+s=</DigestValue></Reference></SignedInfo>
+            <SignatureValue>fHU3XvGbsX4uSc8UypOSG3Xlku6lsCMVM9/0YZDi3eIY20qoRuXko0ZMsuuvGuIA4jGDhOcD5A0NXtbMmML78eHOa7RRJmIe0P3bPyvoz///YvFJHBxpal3pZ/nfnUx4M3wEVmoKWo8LqPUm0RJufIYdMiI01DuWVWkjUNE2KIhilWXO5eaUfnIkVHzjTnDiRGhWcApRi3Rsxw7UkDdGjd5xjPMAvkoG+edIkm4SjwHhlZNpUEbV96lNpDXB3dl5s41RpxiX+3SHMQDuBaPifJV9fHWwDgVYIsZlQpPLPh4AmrkTBVXN/cEK5kDaT+vmofrQGuovsvUiS3K4ZOJEnw==</SignatureValue>
             <KeyInfo>
               <X509Data>
-                <X509Certificate>MIIEGjCCAwKgAwIBAgIRAOujL40uoT6hVFwL1yFBLP4wDQYJKoZIhvcNAQELBQAwRTELMAkGA1UEBhMCR0IxDDAKBgNVBAoTA25oczELMAkGA1UECxMCQ0ExGzAZBgNVBAMTEk5IUyBJTlQgU2lnbmluZyBHMjAeFw0yNDA3MTExMjQ4MDZaFw0yNzA3MTExMzE4MDZaMEcxDDAKBgNVBAoTA25oczEPMA0GA1UECxMGUGVvcGxlMSYwJAYDVQQDDB01NTUyNTQ2MDAxMDZfQXJ2aW5kX05pamFtcHVyZTCBnzANBgkqhkiG9w0BAQEFAAOBjQAwgYkCgYEAv2C3uC4EXq4E2ElA4MnXlT/3U52i9tR2X1xX1VeH0UUSNA9cz+kCFiK8oCC0gG9A4ChaYJE9pVDqAkkQuw+OtLecWYSFDkjuhbIESbWlbREKZLb/D5aDTyY1pvx20OPSWElurXll6m53uPZxeoQI0pQtHnIS986HmCsyg6o6C7cCAwEAAaOCAYUwggGBMA4GA1UdDwEB/wQEAwIGQDBBBgNVHSABAf8ENzA1MDMGCyqGOgCJe2YAAwIAMCQwIgYIKwYBBQUHAgEWFmh0dHBzOi8vcGtpLm5ocy51ay9DUFMwbwYIKwYBBQUHAQEEYzBhMCMGCCsGAQUFBzABhhdodHRwOi8vb2NzcC5uaHMudWsvb2NzcDA6BggrBgEFBQcwAoYuaHR0cDovL3BraS5uaHMudWsvaW50L0cyL3NpZ24vTkhTSU5UU2lnbkcyLmNydDBDBgNVHR8EPDA6MDigNqA0hjJodHRwOi8vY3JsLm5ocy51ay9pbnQvZzIvc2lnbi9OSFNJTlRzaWduRzJfYzE1LmNybDArBgNVHRAEJDAigA8yMDI0MDcxMTEyNDgwNlqBDzIwMjcwNzExMTMxODA2WjAfBgNVHSMEGDAWgBSC+0pvCXGNKbY4bF/C3S/Ou3pu3DAdBgNVHQ4EFgQUpuv+xIZsbV/8EPI3yJz2/O2DWJ8wCQYDVR0TBAIwADANBgkqhkiG9w0BAQsFAAOCAQEAjO1B/+VDzszJAUJMF9FFJ/uxsE9m55RmaLPVtWG55DQYMYt8ENrGfWp5GWzK5lCAdSTSRXUikt+4sR3H4HMiLLvGbVmjBeaat9akMb/8v3ygibDlV9hfMbUiLxP7Fd8hvAPzyGsb6o3Q6h2XBGZJFkDFRPsEw/4doSVtQYpLP4u9ZMuF5CSDOn78M8k6HjUkVHaXWfHfiJWsD7qfiqJEe1LMBrWeaLtnSruf+5yYVh3KR+2dZuy0GIeAoGVStIIuRPLHuBl/3A+75m9YYWTjY4uLPzMaPfaJ1TuxdtpzWf/slQnEy3z/7RGeIG9XZLnFFgXMpU4KPXUm0VD71NHcvA==</X509Certificate>
+                <X509Certificate>MIIEMTCCAxmgAwIBAgIEYCABITANBgkqhkiG9w0BAQsFADA2MQwwCgYDVQQKEwNuaHMxCzAJBgNVBAsTAkNBMRkwFwYDVQQDExBOSFMgSU5UIExldmVsIDJEMB4XDTI1MDEwODEwMjk0OVoXDTI2MDIwMzE4NDYyNFowNjEMMAoGA1UEChMDbmhzMQ8wDQYDVQQLEwZwZW9wbGUxFTATBgNVBAMTDDU1NTMwMTQ5MjEwNjCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALW6Nla54C0AhIuLkFbGrnsavua3E/ADKQCatD79tq6NJmVFbs641Q7WopONlGS8doqK+YS1v1iiNpuITlirktBTJtLYpNpuBGW79qzL/dx7aR5oYi9o/llPu0hbekU+FPjvt3CYaqtsYLzS5kGDlZ8O0Lh8U/zM9TsYZcerxRKWbLRUXlzPUOcL1pN4T4Is+i8J7Z5JxpyYRDZzXxShd9fFJyZpjlHGriRlNn+3YrAsdSTYzhGRDIS5IcUcTbtzNmSEdsBz/RH+9VYBlHmQFa2nqfuGUvdjeqdxj2i1dIUHvNo+B5O8ncXG3/ILMV3BOyZb/PircWPT2C/2LKsud+UCAwEAAaOCAUUwggFBMA4GA1UdDwEB/wQEAwIGQDBlBgNVHSABAf8EWzBZMFcGCyqGOgCJe2YAAwIAMEgwRgYIKwYBBQUHAgEWOmh0dHBzOi8vcGtpLm5ocy51ay9jZXJ0aWZpY2F0ZV9wb2xpY2llcy9jb250ZW50X2NvbW1pdG1lbnQwNQYDVR0fBC4wLDAqoCigJoYkaHR0cDovL2NybC5uaHMudWsvaW50LzJkL2NybGM0NDUuY3JsMCsGA1UdEAQkMCKADzIwMjUwMTA4MTAyOTQ5WoEPMjAyNjAyMDMxODQ2MjRaMB8GA1UdIwQYMBaAFHbQEwruMwOC9dY/FlUs5CGsLqvOMB0GA1UdDgQWBBTqTX39RefMrLvMYwB2tEEvhTKlOzAJBgNVHRMEAjAAMBkGCSqGSIb2fQdBAAQMMAobBFY4LjMDAgSwMA0GCSqGSIb3DQEBCwUAA4IBAQAfdKJ59etOAiJA6QMpzizr76qQU7W/CkCkYRp60uXPUP2M5MWW2WjpZpHfkt6grQKQGE9vlsN6QtbGnhlOs184iaIiaYzXC3fHauEKh4m1c4viy/lL5NrfhNO+e+QUQ1KUdwRF68KgIjEtIRBvPeylvwJ9WERu3LLEeMMaUIgiSDVlM4xTOjDGJNWoRCLE5nNuUTuXQiJzvNkn/3nHpNVFsuh6euUJF38NglRX1iQLoGA/m/pyBkLo3qpICd6JTvlzdiW9s14E+CzFHDL/X2KgByQ1pBJf7ZGagHkg3v2psSZmkthkylEozDmO0XUC80HqPMUTKNNXfVuUFxBKMiMN</X509Certificate>
               </X509Data>
             </KeyInfo>
           </Signature>
@@ -98,7 +98,7 @@
         </AgentPerson>
       </author>
       <legalAuthenticator typeCode="LA" contextControlCode="OP">
-        <time xmlns:hl7="urn:hl7-org:v3" value="20250806125914"/>
+        <time xmlns:hl7="urn:hl7-org:v3" value="20250812120729"/>
         <signatureText nullFlavor="NA"/>
         <AgentPerson classCode="AGNT">
           <id root="1.2.826.0.1285.0.2.0.67" extension="103142999989"/>
