+++ packages/coordinator/tests/resources/test-resources.ts
@@ -353,8 +353,8 @@ export const parentPrescriptions = {
   validSignature: readXml(
     fs.readFileSync(path.join(__dirname, "./signed-prescriptions/ValidSignature.xml"), "utf-8")
   ) as hl7V3.ParentPrescriptionRoot,
-  jceSignature: readXml(
-    fs.readFileSync(path.join(__dirname, "./signed-prescriptions/jceValidSignature.xml"), "utf-8")
+  generateHashSignature: readXml(
+    fs.readFileSync(path.join(__dirname, "./signed-prescriptions/GenerateHashSignature.xml"), "utf-8")
   ) as hl7V3.ParentPrescriptionRoot,
   invalidSignature: readXml(
     fs.readFileSync(path.join(__dirname, "./signed-prescriptions/SignatureIsInvalid.xml"), "utf-8")
